[006] Divided classes into separated files.
