+++ Combat/Include/Ultimate.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include <iostream>
+
+class Ultimate
+{
+  public:
+     virtual std::string useUlt() = 0;
+};
