+++ Combat/Source/WarriorRaceUltimate.cpp
@@ -1,15 +1,7 @@
-#include "Ultimate.hpp"
+#include "WarriorRaceUltimate.hpp"
 #include <iostream>
 
-Ultimate_Mage_Race::Ultimate_Mage_Race(int h, int d) : heal(h), dmg(d){
-}
-
 Ultimate_Warrior_Race::Ultimate_Warrior_Race(int h, int d): heal(h), dmg(d){
-}
-
-std::string Ultimate_Mage_Race::useUlt()
-{
-  return "useMageUlt";
 }
 
 std::string Ultimate_Warrior_Race::useUlt()
