+++ Combat/Test/UltimateTests.cpp
@@ -1,0 +1,4 @@
+#include "Ultimate.hpp"
+#include <gtest/gtest.h>
+
+using namespace ::testing;
